# lab2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Приют и передача животных в семью",
    "description": "REST API приюта: животные, вольеры, заявки на усыновление, акты передачи, пользователи. Даты и время проставляет сервер, в запросах не передаются.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/animals": {
      "get": {
        "tags": [
          "Животные"
        ],
        "summary": "Список животных",
        "operationId": "listAnimals",
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimalResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Животные"
        ],
        "summary": "Добавить животное",
        "operationId": "createAnimal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "201": {
            "description": "Ресурс создан",
            "headers": {
              "Location": {
                "description": "Адрес созданного ресурса",
                "style": "simple"
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/adoption-applications": {
      "get": {
        "tags": [
          "Заявки на усыновление"
        ],
        "summary": "Список заявок",
        "operationId": "listAdoptionApplications",
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdoptionApplicationResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Заявки на усыновление"
        ],
        "summary": "Подать заявку на усыновление",
        "operationId": "createAdoptionApplication",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdoptionApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "201": {
            "description": "Ресурс создан",
            "headers": {
              "Location": {
                "description": "Адрес созданного ресурса",
                "style": "simple"
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdoptionApplicationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/animals/{id}": {
      "get": {
        "tags": [
          "Животные"
        ],
        "summary": "Животное по идентификатору",
        "operationId": "getAnimal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Животные"
        ],
        "summary": "Удалить животное",
        "operationId": "deleteAnimal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "patch": {
        "tags": [
          "Животные"
        ],
        "summary": "Частично обновить животное",
        "operationId": "updateAnimal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimalPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/adoption-applications/{id}": {
      "get": {
        "tags": [
          "Заявки на усыновление"
        ],
        "summary": "Заявка по идентификатору",
        "operationId": "getAdoptionApplication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdoptionApplicationResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Заявки на усыновление"
        ],
        "summary": "Удалить заявку",
        "operationId": "deleteAdoptionApplication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "patch": {
        "tags": [
          "Заявки на усыновление"
        ],
        "summary": "Изменить статус заявки",
        "operationId": "updateAdoptionApplication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdoptionApplicationPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdoptionApplicationResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "Тело ответа с ошибкой",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Момент формирования ответа",
            "example": "2026-09-26T07:07:27.667885Z"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "Код ответа",
            "example": 409
          },
          "error": {
            "type": "string",
            "description": "Причина, соответствующая коду",
            "example": "Conflict"
          },
          "message": {
            "type": "string",
            "description": "Описание причины ошибки"
          },
          "path": {
            "type": "string",
            "description": "Путь запроса",
            "example": "/api/animals/1"
          }
        }
      },
...
      "AnimalRequest": {
        "type": "object",
        "description": "Новое животное приюта",
        "properties": {
          "name": {
            "type": "string",
            "description": "Кличка",
            "example": "Бобик",
            "minLength": 1
          },
          "species": {
            "type": "string",
            "description": "Вид животного",
            "example": "Собака",
            "minLength": 1
          },
          "age": {
            "type": "integer",
            "format": "int32",
            "description": "Возраст в полных годах",
            "example": 3
          },
          "status": {
            "type": "string",
            "description": "Статус размещения, по умолчанию AVAILABLE",
            "enum": [
              "AVAILABLE",
              "RESERVED",
              "HANDED_OVER"
            ],
            "example": "AVAILABLE"
          }
        },
        "required": [
          "age",
          "name",
          "species"
        ]
      },
      "AnimalResponse": {
        "type": "object",
        "description": "Животное приюта",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Идентификатор",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Кличка",
            "example": "Бобик"
          },
          "species": {
            "type": "string",
            "description": "Вид животного",
            "example": "Собака"
          },
          "age": {
            "type": "integer",
            "format": "int32",
            "description": "Возраст в полных годах",
            "example": 3
          },
          "status": {
            "type": "string",
            "description": "Статус размещения",
            "enum": [
              "AVAILABLE",
              "RESERVED",
              "HANDED_OVER"
            ],
            "example": "AVAILABLE"
          }
        }
      },
...
      "AdoptionApplicationRequest": {
        "type": "object",
        "description": "Новая заявка на усыновление",
        "properties": {
          "animalId": {
            "type": "integer",
            "format": "int64",
            "description": "Животное, на которое подана заявка",
            "example": 1
          },
          "applicantId": {
            "type": "integer",
            "format": "int64",
            "description": "Заявитель",
            "example": 1
          },
          "status": {
            "type": "string",
            "description": "Статус заявки, по умолчанию PENDING",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "CANCELLED"
            ],
            "example": "PENDING"
          }
        },
        "required": [
          "animalId",
          "applicantId"
        ]
      },
      "AdoptionApplicationResponse": {
        "type": "object",
        "description": "Заявка на усыновление",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Идентификатор",
            "example": 1
          },
          "animalId": {
            "type": "integer",
            "format": "int64",
            "description": "Животное, на которое подана заявка",
            "example": 1
          },
          "applicantId": {
            "type": "integer",
            "format": "int64",
            "description": "Заявитель",
            "example": 1
          },
          "status": {
            "type": "string",
            "description": "Статус заявки",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "CANCELLED"
            ],
            "example": "PENDING"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Дата и время подачи заявки, проставляются сервером",
            "example": "2026-09-26T07:07:27.613832Z"
          }
        }
      },
...
      "AnimalPatch": {
        "type": "object",
        "description": "Поля животного для изменения, отсутствующие поля не меняются",
        "properties": {
          "name": {
            "type": "string",
            "description": "Кличка",
            "example": "Бобик",
            "pattern": "\\S+"
          },
          "species": {
            "type": "string",
            "description": "Вид животного",
            "example": "Собака",
            "pattern": "\\S+"
          },
          "age": {
            "type": "integer",
            "format": "int32",
            "description": "Возраст в полных годах",
            "example": 4
          },
          "status": {
            "type": "string",
            "description": "Статус размещения",
            "enum": [
              "AVAILABLE",
              "RESERVED",
              "HANDED_OVER"
            ],
            "example": "AVAILABLE"
          }
        }
      },
...
      "AdoptionApplicationPatch": {
        "type": "object",
        "description": "Новый статус заявки",
        "properties": {
          "status": {
            "type": "string",
            "description": "Статус заявки",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "CANCELLED"
            ],
            "example": "APPROVED"
          }
        }
      }
    }
  }
}